feat: added functionality to get the page blocks and block content
