+++ src/scripts/getAnton.ts
@@ -46,23 +46,6 @@ export async function getAnton(blocks : block[]): Promise<string> {
       full_text += "\n";
     }
 
-  //   const filePath = './src/pages/posts/members.md';
-  //
-  //   const extra = "---\ntitle: 'Taken from Notion'\nauthor: 'You'\n---"
-  //
-  //   // layout: ../../layouts/MarkdownPostLayout.astro\n
-  //
-  //   // Write the Markdown content to the file
-  //   fs.writeFile(filePath, extra + "  \n" + full_text, (err) => {
-  //     if (err) {
-  //       console.error('Error writing file:', err);
-  //     } else {
-  //       console.log('Markdown file written successfully.');
-  //     }
-  //   });
-  //
-  // })();
-
   return full_text;
 
 }
