+++ src/scripts/getDetails.ts
@@ -2,6 +2,7 @@ import { Client } from "@notionhq/client";
 import * as fs from "fs";
 import { type block } from "../types/block";
 import {getAnton} from "./getAnton.ts";
+import {fetchPageBlocks} from "./fetchPageBlocks.ts";
 
 
 
@@ -48,21 +49,4 @@ export async function getDetails(): Promise<void> {
       console.error('Error querying database:', error);
   }
 
-}
-
-async function fetchPageBlocks(notion : Client, blockID: string) : Promise<block[]>{
-  const blocks: block[] = [];
-  let cursor: string | undefined = undefined;
-
-  do {
-    const response = await notion.blocks.children.list({
-      block_id: blockID,
-      start_cursor: cursor,
-      page_size: 50,
-    });
-    blocks.push(...response.results as block[]);
-    cursor = response.next_cursor ?? undefined;
-  } while (cursor);
-
-  return blocks;
 }
